Remove DBFlow in favor of Realm for library persistence

+++ app/src/main/java/net/sigmabeta/chipbox/ui/settings/SettingsActivity.kt
@@ -44,7 +44,7 @@ class SettingsActivity : BaseActivity(), SettingsView, ItemListView<VoiceViewHol
      * ItemListView
      */
 
-    override fun onItemClick(position: Long, clickedViewHolder: VoiceViewHolder) {
+    override fun onItemClick(position: Int, clickedViewHolder: VoiceViewHolder) {
         presenter.onItemClick(position)
     }
